requirements.txt
openai
langchain
faiss-cpu
pandas
streamlit
python-dotenv